+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - (DEV) Fixed the `gulp README-to-local` and `gulp README-to-online`.
 - (DEV) Fixed code examples on JSDocs for `gulp build-docs`.
 - (DEV) Fixed the tests that had errors.
+- Fixes the chart legend color for `line` charts.
 ### Changed
 - (DEV) `npm run test-all-serial` runs all the tests in serial.
 ### Added